+++ src/circuit/mod.rs
@@ -1,6 +1,4 @@
 use crate::tensor::*;
-/// Utility functions for building gates.
-pub mod chip;
 /// Element-wise operations using lookup tables.
 pub mod eltwise;
 /// Structs and methods for configuring and assigning many operations to a gate within a Halo2 circuit.
